% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cjs_codes.R
\name{get_cjs_codes}
\alias{get_cjs_codes}
\title{Get Criminal Justice Codes}
\usage{
get_cjs_codes(
  u = paste0("https://www.criminaljusticehub.org.uk/wp-content/uploads/",
    "2022/10/cjs-offence-index-sept-2022.ods")
)
}
\arguments{
\item{u}{The URL of the Criminal Justice Hub website.}
}
\description{
This function gets the criminal justice codes from the Criminal Justice Hub website.
See \href{https://www.criminaljusticehub.org.uk/jargon-buster/cjs-offence-code/}{www.criminaljusticehub.org.uk}
}
\examples{
get_cjs_codes()
}
